Apply multiple fixes to new panel types

+++ src/components/panels/Functions/sum.tsx
@@ -63,10 +63,10 @@ const create = (panelId: number): Panel => {
 
         const allNumbers = eps.reduce((a, ep) => a && !isNaN(values[ep]), true);
 
-        console.log('execute sum', panel.addendEps, eps.reduce((a, ep) => a + parseInt(values[ep]), 0));
+        console.log('execute sum', panel.addendEps, eps.reduce((a, ep) => a + parseFloat(values[ep]), 0));
 
         if (!allNumbers) return { outputResult: '' };
-        return { outputResult: eps.reduce((a, ep) => a + parseInt(values[ep]), 0) };
+        return { outputResult: eps.reduce((a, ep) => a + parseFloat(values[ep]), 0) };
     };
 
     return {
